tests: create libunicorn.so.0 symlink
This hack allows the unit tests to link against libunicorn, after
the SONAME was added in 4860fdb3.

Becuase libunicorn.so has an SONAME of libunicorn.so.0, the linker uses
the SONAME for the DT_NEEDED entry of the tests. But because a library
with that name does not exist, they would fail to run. This symlink
works around that issue, without touching the toplevel Makefile.