fix install logic after changing LIBRARY
Now that LIBRARY is the versioned name (e.g. "libunicorn.so.0"), we
don't need this rename step; just create the symlink for libunicorn.so.
If VERSION_EXT isn't set, then the library we installed already has the
correct name, and no symlink is necessary.

We should probably be running 'ldconfig' to handle this, or better yet,
using libtool.